+++ backend/src/main/java/coogether/backend/domain/Report.java
@@ -26,7 +26,7 @@ public class Report {
     @Id
     @GeneratedValue
     @Column(name = "report_id", nullable = false)
-    private int reportId;
+    private Long reportId;
 
     @JsonIgnore
     @ManyToOne(fetch = LAZY)
